feat(VkWrapper): implement VkWrapper systems and enhance functionality
